requirements.txt
streamlit
scikit-learn
pandas
numpy
pillow
xlsxwriter
